requirements.txt
flask
langchain
langchain-groq
langchain-community
langchain-core
langchain-huggingface
faiss-cpu
python-dotenv
sentence-transformers
numpy
pytesseract
Pillow
deep-translator
langdetect
groq
requests
